pyproject.toml
[project]
name = "aulas"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "flask>=3.1.1",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.8",
]
